+++ projects/mods/roll-protagonist/RollProtagonist.Frontend/ModEntry.cs
@@ -1,3 +1,4 @@
+using RollProtagonist.Frontend.NewGamePlus.Patching;
 using TaiwuModdingLib.Core.Plugin;
 
 namespace RollProtagonist.Frontend;
